Fixes and style changes matching @kuldar's design specs. Details on https://github.com/prisma/graphql-playground/pull/897 comments
